core animation functionality refactored
