+++ ptplot/ptplot.py
@@ -127,55 +127,11 @@ class PTPlot:
             figure_object.legend.click_policy = "mute"
             figures.append(figure_object)
 
-        widgets = []
+        plot_grid = gridplot(figures, ncols=self.facet_layer.num_col)
         # TODO: could this be handled by using bokeh's tagging functionality?
         # Probably could, by storing the closure with the plot
         if self.animation_layer is not None:
-            min_frame = mapping_data[self.animation_layer.frame_mapping].min()
-            max_frame = mapping_data[self.animation_layer.frame_mapping].max()
-            play_pause = Toggle(label="► Play", active=False)
-            widgets.append(play_pause)
-            slider = Slider(start=min_frame, end=max_frame, value=min_frame, step=1, title="Frame")
-            play_pause_js = CustomJS(
-                args={"slider": slider, "min_frame": min_frame, "max_frame": max_frame},
-                code="""
-var check_and_iterate = function(){
-    var slider_val = slider.value;
-    var toggle_val = cb_obj.active;
-    if(toggle_val == false) {
-        cb_obj.label = '► Play';
-        clearInterval(play_pause_loop);
-        }
-    else if(slider_val == max_frame) {
-        cb_obj.label = '► Play';
-        slider.value = min_frame;
-        cb_obj.active = false;
-        clearInterval(play_pause_loop);
-        }
-    else if(slider_val !== max_frame){
-        slider.value = slider_val + 1;
-        }
-    else {
-    clearInterval(play_pause_loop);
-        }
-}
-if(cb_obj.active == false){
-    cb_obj.label = '► Play';
-    clearInterval(play_pause_loop);
-}
-else {
-    cb_obj.label = '❚❚ Pause';
-    var play_pause_loop = setInterval(check_and_iterate, 100);
-};
-                """,
-            )
-            play_pause.js_on_change("active", play_pause_js)
-            widgets.append(slider)
-            for animation in animations:
-                callback = animation(self.animation_layer.frame_mapping, min_frame)
-                slider.js_on_change("value", callback)
-        plot_grid = gridplot(figures, ncols=self.facet_layer.num_col)
-        if len(widgets) > 0:
+            widgets = self.animation_layer.animate(mapping_data, animations)
             plot_grid.children.append(row(widgets))
         return plot_grid
